Close #24 - Add putStr and putErrStr to ConsoleEffect
