+++ scalaz-effect/src/main/scala/effectie/scalaz/ConsoleEffect.scala
@@ -1,9 +1,10 @@
 package effectie.scalaz
 
-import effectie.YesNo
-
 import scalaz._
 import Scalaz._
+
+import effectie.ConsoleEffect.ConsoleEffectWithoutBind
+import effectie.YesNo
 
 
 trait ConsoleEffect[F[_]] extends effectie.ConsoleEffect[F]
@@ -14,15 +15,9 @@ object ConsoleEffect {
   implicit def consoleEffectF[F[_] : EffectConstructor : Bind]: ConsoleEffect[F] =
     new ConsoleEffectF[F]
 
-  final class ConsoleEffectF[F[_] : EffectConstructor : Bind] extends ConsoleEffect[F] {
-    override def readLn: F[String] =
-      EffectConstructor[F].effectOf(scala.io.StdIn.readLine)
-
-    override def putStrLn(value: String): F[Unit] =
-      EffectConstructor[F].effectOf(Console.out.println(value))
-
-    override def putErrStrLn(value: String): F[Unit] =
-      EffectConstructor[F].effectOf(Console.err.println(value))
+  final class ConsoleEffectF[F[_] : EffectConstructor : Bind]
+    extends ConsoleEffectWithoutBind[F]
+    with ConsoleEffect[F] {
 
     @SuppressWarnings(Array("org.wartremover.warts.Recursion"))
     override def readYesNo(prompt: String): F[YesNo] = for {
